# _oas/openapi.yaml
openapi: 3.0.3
info:
  title: Go Faster Github Archive Coordinator
  description: Coordinates Archive Workers
  version: 0.1.0
servers:
  - url: 'https://localhost:8080'
paths:
  /progress:
    post:
      operationId: progress
      description: report progress
      responses:
        200:
          description: Status
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Status"
        default:
          $ref: "#/components/responses/Error"
      parameters:
        - $ref: "#/components/parameters/Token"
      requestBody:
        required: true

        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/Progress"

  /status:
    get:
      operationId: "status"
      description: "get status"
      responses:
        200:
          description: Coordinator status
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Status"
        default:
          $ref: "#/components/responses/Error"
  /job/poll:
    post:
      operationId: "poll"
      description: Request job from coordinator
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        200:
          description: Job is available
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Job"
        default:
          $ref: "#/components/responses/Error"
components:
  parameters:
    Token:
      in: header
      required: true
      name: X-Token
      description: Worker token
      schema:
        type: string
  responses:
    Error:
      description: "Error while processing request"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Progress:
      type: object
      required:
        - done
        - key
      properties:
        done:
          description: ok if completed
          type: boolean
        key:
          description: chunk key
          type: string
          minLength: 5
        size_bytes:
          type: integer
          format: int64
        ready_bytes:
          type: integer
          format: int64
        sha256_input:
          type: string
          minLength: 64
          maxLength: 64
        sha256_content:
          type: string
          minLength: 64
          maxLength: 64
        sha256_output:
          type: string
          minLength: 64
          maxLength: 64
    Status:
      type: object
      properties:
        message:
          type: string
      required: [ message ]

    Error:
      type: object
      description: "error description"
      properties:
        message:
          type: string
      required: [ message ]

    JobNothing:
      type: object
      description: Do nothing
      required:
        - type
      properties:
        type:
          type: string

    JobDownload:
      type: object
      description: Download chunk
      required:
        - type
        - date
      properties:
        type:
          type: string
        date:
          type: string
          format: datetime

    Job:
      type: object
      description: job to perform on worker
      discriminator:
        propertyName: "type"
        mapping:
          nothing: JobNothing
          download: JobDownload
      oneOf:
        - $ref: "#/components/schemas/JobNothing"
        - $ref: "#/components/schemas/JobDownload"